consolidate code into cli discovery file
