+++ extension/src/setup.ts
@@ -4,30 +4,12 @@ import {
   quickPickValue,
   quickPickYesOrNo
 } from './vscode/quickPick'
-import {
-  ConfigKey,
-  getConfigValue,
-  setConfigValue,
-  setUserConfigValue
-} from './vscode/config'
+import { ConfigKey, setConfigValue } from './vscode/config'
 import { pickFile } from './vscode/resourcePicker'
 import { getFirstWorkspaceFolder } from './vscode/workspaceFolders'
-import { Response } from './vscode/response'
 import { getSelectTitle, Title } from './vscode/title'
-import { Toast } from './vscode/toast'
-import {
-  getPythonBinPath,
-  isPythonExtensionInstalled,
-  selectPythonInterpreter
-} from './extensions/python'
-import {
-  CliCompatible,
-  isCliCompatible,
-  warnVersionIncompatible,
-  isVersionCompatible,
-  warnAheadOfLatestTested,
-  warnUnableToVerifyVersion
-} from './cli/dvc/version'
+import { isPythonExtensionInstalled } from './extensions/python'
+import { extensionCanRunCli } from './cli/dvc/discovery'
 
 const setConfigPath = async (
   option: ConfigKey,
@@ -169,181 +151,6 @@ export const setupWorkspace = async (): Promise<boolean> => {
   }
 
   return pickCliPath()
-}
-
-const getToastOptions = (isPythonExtensionInstalled: boolean): Response[] => {
-  return isPythonExtensionInstalled
-    ? [Response.SETUP_WORKSPACE, Response.SELECT_INTERPRETER, Response.NEVER]
-    : [Response.SETUP_WORKSPACE, Response.NEVER]
-}
-
-const warnUserCLIInaccessible = async (
-  extension: IExtension,
-  isMsPythonInstalled: boolean,
-  warningText: string
-): Promise<void> => {
-  if (getConfigValue<boolean>(ConfigKey.DO_NOT_SHOW_CLI_UNAVAILABLE)) {
-    return
-  }
-
-  const response = await Toast.warnWithOptions(
-    warningText,
-    ...getToastOptions(isMsPythonInstalled)
-  )
-
-  switch (response) {
-    case Response.SELECT_INTERPRETER:
-      return selectPythonInterpreter()
-    case Response.SETUP_WORKSPACE:
-      return extension.setupWorkspace()
-    case Response.NEVER:
-      return setUserConfigValue(ConfigKey.DO_NOT_SHOW_CLI_UNAVAILABLE, true)
-  }
-}
-
-const warnUserCLIInaccessibleAnywhere = async (
-  extension: IExtension,
-  globalDvcVersion: string | undefined
-): Promise<void> => {
-  const binPath = await getPythonBinPath()
-
-  return warnUserCLIInaccessible(
-    extension,
-    true,
-    `The extension is unable to initialize. The CLI was not located using the interpreter provided by the Python extension. ${
-      globalDvcVersion ? globalDvcVersion + ' is' : 'The CLI is also not'
-    } installed globally. For auto Python environment activation, ensure the correct interpreter is set. Active Python interpreter: ${binPath}.`
-  )
-}
-
-const warnUser = (
-  extension: IExtension,
-  cliCompatible: CliCompatible,
-  version: string | undefined
-): void => {
-  if (!extension.hasRoots()) {
-    return
-  }
-  switch (cliCompatible) {
-    case CliCompatible.NO_BEHIND_MIN_VERSION:
-      return warnVersionIncompatible(version as string, 'CLI')
-    case CliCompatible.NO_CANNOT_VERIFY:
-      warnUnableToVerifyVersion()
-      return
-    case CliCompatible.NO_MAJOR_VERSION_AHEAD:
-      return warnVersionIncompatible(version as string, 'extension')
-    case CliCompatible.NO_NOT_FOUND:
-      warnUserCLIInaccessible(
-        extension,
-        isPythonExtensionInstalled(),
-        'An error was thrown when trying to access the CLI.'
-      )
-      return
-    case CliCompatible.YES_MINOR_VERSION_AHEAD_OF_TESTED:
-      return warnAheadOfLatestTested()
-  }
-}
-
-type CanRunCli = {
-  isAvailable: boolean
-  isCompatible: boolean | undefined
-}
-
-const getVersionDetails = async (
-  extension: IExtension,
-  cwd: string,
-  tryGlobalCli?: true
-): Promise<
-  CanRunCli & {
-    cliCompatible: CliCompatible
-    version: string | undefined
-  }
-> => {
-  const version = await extension.getCliVersion(cwd, tryGlobalCli)
-  const cliCompatible = isVersionCompatible(version)
-  const isCompatible = isCliCompatible(cliCompatible)
-  return { cliCompatible, isAvailable: !!isCompatible, isCompatible, version }
-}
-
-const processVersionDetails = (
-  extension: IExtension,
-  cliCompatible: CliCompatible,
-  version: string | undefined,
-  isAvailable: boolean,
-  isCompatible: boolean | undefined
-): CanRunCli => {
-  warnUser(extension, cliCompatible, version)
-  return {
-    isAvailable,
-    isCompatible
-  }
-}
-
-const tryGlobalFallbackVersion = async (
-  extension: IExtension,
-  cwd: string
-): Promise<CanRunCli> => {
-  const tryGlobal = await getVersionDetails(extension, cwd, true)
-  const { cliCompatible, isAvailable, isCompatible, version } = tryGlobal
-
-  if (extension.hasRoots() && !isCompatible) {
-    warnUserCLIInaccessibleAnywhere(extension, version)
-  }
-  if (
-    extension.hasRoots() &&
-    cliCompatible === CliCompatible.YES_MINOR_VERSION_AHEAD_OF_TESTED
-  ) {
-    warnAheadOfLatestTested()
-  }
-
-  if (isCompatible) {
-    extension.unsetPythonBinPath()
-  }
-
-  return { isAvailable, isCompatible }
-}
-
-const extensionCanAutoRunCli = async (
-  extension: IExtension,
-  cwd: string
-): Promise<CanRunCli> => {
-  const {
-    cliCompatible: pythonCliCompatible,
-    isAvailable: pythonVersionIsAvailable,
-    isCompatible: pythonVersionIsCompatible,
-    version: pythonVersion
-  } = await getVersionDetails(extension, cwd)
-
-  if (pythonCliCompatible === CliCompatible.NO_NOT_FOUND) {
-    return tryGlobalFallbackVersion(extension, cwd)
-  }
-  return processVersionDetails(
-    extension,
-    pythonCliCompatible,
-    pythonVersion,
-    pythonVersionIsAvailable,
-    pythonVersionIsCompatible
-  )
-}
-
-const extensionCanRunCli = async (
-  extension: IExtension,
-  cwd: string
-): Promise<CanRunCli> => {
-  if (await extension.isPythonExtensionUsed()) {
-    return extensionCanAutoRunCli(extension, cwd)
-  }
-
-  const { cliCompatible, isAvailable, isCompatible, version } =
-    await getVersionDetails(extension, cwd)
-
-  return processVersionDetails(
-    extension,
-    cliCompatible,
-    version,
-    isAvailable,
-    isCompatible
-  )
 }
 
 export const setup = async (extension: IExtension) => {
